Create dedicated runtime exception package
Until bulk operation support was added recently, all exception classes
defined in the SCIM SDK were checked exceptions. With Jackson 3
support, at least one more unchecked exception will be added. To
highlight this behavior, a dedicated "runtime" package has been added.

Reviewer: dougbulkley
Reviewer: vyhhuang
