+++ scim2-sdk-common/src/main/java/com/unboundid/scim2/common/bulk/BulkRequest.java
@@ -41,8 +41,8 @@ import com.unboundid.scim2.common.annotations.Attribute;
 import com.unboundid.scim2.common.annotations.NotNull;
 import com.unboundid.scim2.common.annotations.Nullable;
 import com.unboundid.scim2.common.annotations.Schema;
-import com.unboundid.scim2.common.exceptions.BulkRequestException;
-import com.unboundid.scim2.common.exceptions.BulkResponseException;
+import com.unboundid.scim2.common.exceptions.runtime.BulkRequestException;
+import com.unboundid.scim2.common.exceptions.runtime.BulkResponseException;
 import com.unboundid.scim2.common.exceptions.ContentTooLargeException;
 import com.unboundid.scim2.common.exceptions.RateLimitException;
 import com.unboundid.scim2.common.exceptions.ResourceConflictException;
